refactor: pull out non-paint-blocking scheduler from both queue implementations
