+++ packages/runtime/src/dom/queue-sorted-insert.ts
@@ -6,25 +6,9 @@ import {
   getOwnerScope,
   ownerOffset,
 } from "./scope";
+import { schedule } from "./schedule";
 
 type ExecFn = (arg?: any) => void;
-
-const { port1, port2 } = new MessageChannel();
-let queued: boolean;
-
-port1.onmessage = () => {
-  queued = false;
-  run();
-};
-
-function flushAndWaitFrame() {
-  run();
-  requestAnimationFrame(triggerMacroTask);
-}
-
-function triggerMacroTask() {
-  port2.postMessage(0);
-}
 
 let queuedFns: unknown[] = [];
 let queuedNext: unknown[] = [];
@@ -55,10 +39,7 @@ export function queue<T extends ExecFn>(
     queuedFns[index + QueueOffsets.SCOPE] !== scope ||
     queuedFns[index + QueueOffsets.OFFSET] !== offset
   ) {
-    if (!queued) {
-      queued = true;
-      queueMicrotask(flushAndWaitFrame);
-    }
+    schedule();
 
     for (let i = queuedFns.length - 1; i >= index; i--) {
       queuedFns[i + QueueOffsets.TOTAL] = queuedFns[i];
